refactoring; start add move semantics
