+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-project(boost-asio-example)
 cmake_minimum_required(VERSION 3.10)
+project(boost-asio-example)
 
 add_subdirectory(example)
